moved processor to separate struct
